Fix big counts issues

+++ alltoallv/logger.h
@@ -9,6 +9,7 @@
 #include <unistd.h>
 #include <assert.h>
 #include <string.h>
+#include <inttypes.h>
 
 #include "alltoallv_profiler.h"
 
@@ -33,7 +34,7 @@ typedef struct logger
 
 extern logger_t *logger_init();
 extern void logger_fini(logger_t **l);
-extern void log_profiling_data(logger_t *logger, int avCalls, int avCallStart, int avCallsLogged, avSRCountNode_t *counters_list, avTimingsNode_t *times_list);
+extern void log_profiling_data(logger_t *logger, uint64_t avCalls, uint64_t avCallStart, uint64_t avCallsLogged, avSRCountNode_t *counters_list, avTimingsNode_t *times_list);
 extern void log_timing_data(logger_t *logger, avTimingsNode_t *times_list);
 extern int *lookup_rank_counters(int data_size, counts_data_t **data, int rank);
 extern char *compress_int_array(int *array, int size);
